Added checks to final door, mayor office and central town
The final door checks for all 4 WinCards before showing the "point of no return" dialogue.

The trophy case in the mayors office will check for the StampCrd before moving out of the way of the secret passage.

The roadblock in front of the Expo will check for the StampCrd before erasing itself from existence.